Extract dangling entry partial
